# trading-service/docs/swagger.yaml
basePath: /
definitions:
...
  types.Exchange:
    properties:
      acronym:
        type: string
      close_time:
        type: string
      country:
        type: string
      currency:
        type: string
      id:
        type: integer
      mic_code:
        type: string
      name:
        type: string
      open_time:
        type: string
      timezone:
        type: string
    type: object
  types.ForexPair:
    properties:
      base_currency:
        type: string
      exchange_rate:
        type: number
      id:
        type: integer
      liquidity:
        description: '"High", "Medium", "Low"'
        type: string
      listing:
        $ref: '#/definitions/types.Listing'
      listing_id:
        type: integer
      quote_currency:
        type: string
    type: object
...
  types.Listing:
    properties:
      Subtype:
        description: '"ETF", "Common Stock", "Call Option", "Put Option"'
        type: string
      ask:
        type: number
      bid:
        type: number
      contract_size:
        type: integer
      created_at:
        type: string
      exchange:
        $ref: '#/definitions/types.Exchange'
      exchange_id:
        type: integer
      id:
        type: integer
      last_refresh:
        type: string
      lastPrice:
        type: number
      name:
        type: string
      ticker:
        type: string
      type:
        description: '"Stock", "Forex", "Future", "Option"'
        type: string
      updated_at:
        type: string
    type: object
  types.ListingDailyPriceInfo:
    properties:
      availableQuantity:
        type: integer
      change:
        type: number
      date:
        type: string
      high:
        type: number
      id:
        type: integer
      listing:
        $ref: '#/definitions/types.Listing'
      listing_id:
        type: integer
      low:
        type: number
      price:
        type: number
    type: object
...
  types.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Trading Service API
  title: Trading Service
  version: "1.0"
paths:
...
  /forex:
    get:
      description: Vraća listu svih listinga koji predstavljaju forex valutne parove.
      produces:
      - application/json
      responses:
        "200":
          description: Lista svih forex parova
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Listing'
                  type: array
              type: object
        "500":
          description: Interna greška servera pri preuzimanju forex parova
          schema:
            $ref: '#/definitions/types.Response'
      summary: Preuzimanje svih forex parova
      tags:
      - Forex
  /forex/{base}/{quote}:
    get:
      description: Vraća osnovne podatke (Listing) i detalje (ForexPair) za forex
        par specificiran osnovnom i kvotnom valutom.
      parameters:
      - description: Osnovna valuta (ISO kod)
        example: EUR
        in: path
        name: base
        required: true
        type: string
      - description: Kvotna valuta (ISO kod)
        example: USD
        in: path
        name: quote
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detalji forex para
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  properties:
                    details:
                      $ref: '#/definitions/types.ForexPair'
                    listing:
                      $ref: '#/definitions/types.Listing'
                  type: object
              type: object
        "404":
          description: Forex par nije pronađen
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Interna greška servera pri preuzimanju detalja forex para
          schema:
            $ref: '#/definitions/types.Response'
      summary: Preuzimanje forex para po valutama
      tags:
      - Forex
  /forex/{base}/{quote}/history:
    get:
      description: Preuzima dnevne istorijske podatke za forex par u vremenskom opsegu.
      parameters:
      - description: Osnovna valuta
        example: EUR
        in: path
        name: base
        required: true
        type: string
      - description: Kvotna valuta
        example: USD
        in: path
        name: quote
        required: true
        type: string
      - description: Početni datum (YYYY-MM-DD). Podrazumevano pre 30 dana.
        example: "2023-09-27"
        format: date
        in: query
        name: startDate
        type: string
      - description: Krajnji datum (YYYY-MM-DD). Podrazumevano današnji dan.
        example: "2023-10-27"
        format: date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista istorijskih podataka
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.ListingDailyPriceInfo'
                  type: array
              type: object
        "400":
          description: Neispravan format datuma
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Forex par nije pronađen
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Greška pri preuzimanju istorijskih podataka iz baze
          schema:
            $ref: '#/definitions/types.Response'
      summary: Istorijski podaci za forex par u vremenskom opsegu
      tags:
      - Forex
...
securityDefinitions:
  BearerAuth:
    description: 'Unesite token. Primer: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"